folder configs are not read by python
Mac places a ".DS_Store" file in EVERY DIRECTORY. Python tries to read it, but its either `bytes` or does not have 3 lines (required for a haiku). This fix simply gets around this by removing ".DS_Store" from the list of available files.
Windows places a "desktop.ini" file IF you configure the folder. This also has a fix. (Though it would not cause a crash, it would just show [.ShellClassInfo] ect.
 complex